refactor(web): update path of simplepage to the new one
